chore: bring file-size lint green on the unified branch
src/adapt/types.rs (429 → 216): extracted the inline test module to
src/adapt/types_tests.rs and re-included it via #[path = ...] mod tests;.
Same pattern as wizard_fields_tests.rs and issue_completion_tests.rs.

src/tui/app/data_handler.rs (406, 6 over): added an allowlist entry
with a 2026-07-22 deadline and a concrete split plan — extract the
AdaptScan/Analyze/Consolidate/Plan/Scaffold arms into a dedicated
handler module. The dupe-prevention work in #455 added the
IssueAlreadyExists arm that pushed it over.

3117 tests still pass; clippy / fmt clean.
