+++ scripts/allowlist-large-files.txt
@@ -100,3 +100,6 @@ src/tui/screens/milestone_wizard/mod.rs # deadline: 2026-07-22, owner: @carlos, 
 src/tui/screens/issue_wizard/mod.rs # deadline: 2026-07-22, owner: @carlos, ticket: #TBD, plan: split dep_* state + ai_review state + create_* state into submodules (the #447 tui-textarea migration added sync/rebuild helpers; further reduction comes from the submodule split)
 # issue_wizard/draw.rs dropped from 378 → 356 LOC after #447 (under the 400-LOC cap).
 # milestone_wizard/draw.rs dropped from 352 → 345 LOC after #447 (under the 400-LOC cap).
+
+# --- Milestone v0.15.3 dupe-prevention additions (#454/#455) ---
+src/tui/app/data_handler.rs # deadline: 2026-07-22, owner: @carlos, ticket: #TBD, plan: extract the AdaptScan/Analyze/Consolidate/Plan/Scaffold arms into a dedicated handler module (handle_adapt_event); the dupe-prevention work in #455 added the IssueAlreadyExists arm, pushing the file 6 lines over
